+++ changes.md
@@ -3,3 +3,4 @@
 Release for [2014-07-XX]
 
 - changes.md file created;
+- OO_JS use of prototype in models